make stack overflow less likely in state machines
The state machines for bzip2 and inflate decompression inadvertently
assumed that the implementation performs tail call optimizations.
Several popular implementations do not, leading to stack overflows while
attempting to decompress large files.

This change attempts to fix that by only enabling the tail call
mechanism inside the state machine if we are running in an
implementation we can trust to perform tail call optimizations.
Currently, we can trust SBCL and CMUCL; the optimization settings for
the state machine functions have been tweaked to try to enforce tail
call optimization on such implementations.

This change may make performance worse for some
implementations (Lispworks and Allegro come immediately to mind); if
users of chipz on those implementations can perform benchmarks and
report back, I am happy to add them to the list of "trusted"
implementations.